Fix: Brand message not being sent, send lowercase locale, ensure the MCPL default listener comes first
